refactor: moved common package from routers to app folder
